refactor: moved AudioPlayer to AudioContext
