+++ packages/react-native-audio-api/common/cpp/core/Constants.h
@@ -6,13 +6,14 @@
 // https://webaudio.github.io/web-audio-api/
 
 namespace audioapi {
-constexpr int SAMPLE_RATE = 48000;
+constexpr int DEFAULT_SAMPLE_RATE = 48000;
+constexpr int RENDER_QUANTUM_SIZE = 128;
 constexpr int CHANNEL_COUNT = 2;
 
 constexpr float MOST_POSITIVE_SINGLE_FLOAT = static_cast<float>(std::numeric_limits<float>::max());
 constexpr float MOST_NEGATIVE_SINGLE_FLOAT = static_cast<float>(std::numeric_limits<float>::lowest());
 
-constexpr float NYQUIST_FREQUENCY = SAMPLE_RATE / 2.0;
+constexpr float NYQUIST_FREQUENCY = DEFAULT_SAMPLE_RATE / 2.0;
 static float MAX_DETUNE = 1200 * std::log2(MOST_POSITIVE_SINGLE_FLOAT);
 constexpr float MAX_GAIN = MOST_POSITIVE_SINGLE_FLOAT;
 constexpr float MAX_PAN = 1.0;
